semafori kraj

+++ h/Print.hpp
@@ -5,7 +5,7 @@
 
    extern void printString(char const *string);
 
-   extern void printInteger(uint64 integer);
+   extern void printInteger(int integer);
